Renamed 'reducers' to 'reducer'.
