+++ assets/pages/oneoff-contributions/components/contributionsCheckoutContainer.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 
 import ContributionsCheckout from 'components/contributionsCheckout/contributionsCheckout';
 
-import { type PageState as State } from '../oneOffContributionsReducers';
+import { type PageState as State } from '../oneOffContributionsReducer';
 import OneoffContributionsPayment from './oneoffContributionsPayment';
 import OneoffInlineContributionsPayment from './oneoffInlineContributionsPayment';
