+++ assets/pages/regular-contributions/components/contributionsCheckoutContainer.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 
 import ContributionsCheckout from 'components/contributionsCheckout/contributionsCheckout';
 
-import { type PageState as State } from '../regularContributionsReducers';
+import { type PageState as State } from '../regularContributionsReducer';
 import RegularContributionsPayment from './regularContributionsPayment';
 import RegularInlineContributionsPayment from './regularInlineContributionsPayment';
